requirements.txt
openai
python-dotenv
pandas
numpy
tqdm
jsonschema